Extract tests into file
